% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-data.R
\docType{data}
\name{s}
\alias{s}
\title{sample 'msl.trend' object}
\format{msl.trend object}
\usage{
data(s)
}
\description{
Output of call to \code{\link{msl.trend}} used extensively in examples throughout
this Manual.
}
\details{
This \code{\link{msl.trend}} object is used extensively in the
examples throughout this manual in order to call the object direct rather than
producing the same via original code which can be computationally expensive. This
object results from a decomposition of the Baltimore record, filling gaps with
spline interpolation and using 500 iterations to generate error margins via
bootstrapping.

\strong{Note: }Ordinarily the user would call 'File.csv' direct from working
directory, creating the 'msl.trend' object using the following sample code:

s <- msl.trend('Balt.csv', fillgaps = 3, iter = 500, 'BALTIMORE, USA') # DON'T RUN
}
\examples{
data(s)
str(s) # check structure of object
}
\seealso{
\code{\link{msl.trend}}, \code{\link{msl.forecast}},
\code{\link{msl.plot}}, \code{\link{msl.pdf}}, \code{\link{summary}},
\code{\link{Balt}}.
}
\keyword{datasets}

